Custom Hook for authentication and authorization
